Update CREATE e PUT

+++ trabalho4/src/main/java/br/com/pc3/trabalho4/repository/EmpregadoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import br.com.pc3.trabalho4.domain.Empregado;
 
 @Repository
-public interface EmpregadoRepository extends CrudRepository<Empregado, Integer>{
+public interface EmpregadoRepository extends CrudRepository<Empregado, Long>{
 	
 }
